+++ CHANGES.md
@@ -4,6 +4,10 @@ moodle-local_boostnavigation
 Changes
 -------
 
+### Unreleased
+
+* 2019-02-05 - Improve the accessibility of the collapsible nodes, let users open / close the MyCourses node with the keyboard.
+
 ### v3.5-r4
 
 * 2019-02-04 - Fix a flaw in the language pack for the node title placeholders.